Move Renamer and utility functions out of rename.cc
